+++ cogniopenapp/lib/src/database/app_database_seed_data.dart
@@ -1,12 +1,12 @@
 import 'dart:io';
 
+import 'package:cogniopenapp/src/database/controller/photo_controller.dart';
 import 'package:cogniopenapp/src/database/model/audio.dart';
 import 'package:cogniopenapp/src/database/model/photo.dart';
 import 'package:cogniopenapp/src/database/model/video.dart';
 import 'package:cogniopenapp/src/database/repository/audio_repository.dart';
 import 'package:cogniopenapp/src/database/repository/photo_repository.dart';
 import 'package:cogniopenapp/src/database/repository/video_repository.dart';
-import 'package:cogniopenapp/src/media_controller.dart';
 import 'package:cogniopenapp/src/utils/file_manager.dart';
 
 class AppDatabaseSeedData {
@@ -135,7 +135,7 @@ class AppDatabaseSeedData {
       File? photoFile = await FileManager.loadAssetFile(
           'assets/seed_data_files/cat.png', 'cat.png');
       List<String>? tagsList = ['pet', 'cat'];
-      int newPhotoId = await MediaController.addPhoto(
+      int newPhotoId = await PhotoController.addPhoto(
         title: 'Cat',
         description: 'A photo of my pet cat, Kit Kat.',
         tags: tagsList,
