+++ cogniopenapp/lib/src/database/repository/audio_repository.dart
@@ -2,6 +2,17 @@ import 'package:cogniopenapp/src/database/app_database.dart';
 import 'package:cogniopenapp/src/database/model/media.dart';
 import 'package:cogniopenapp/src/database/model/audio.dart';
 
+const String tableAudios = 'audios';
+
+class AudioFields extends MediaFields {
+  static final List<String> values = [
+    ...MediaFields.values,
+    summary,
+  ];
+
+  static const String summary = 'summary';
+}
+
 class AudioRepository {
   static final AudioRepository instance = AudioRepository._init();
