+++ cogniopenapp/test/conversation_test.dart
@@ -1,7 +1,3 @@
-import 'package:cogniopenapp/src/media.dart';
-import 'package:flutter_test/flutter_test.dart';
-import 'package:cogniopenapp/src/conversation.dart';
-
 // TODO -- Fix Tests
 
 /*
