+++ cogniopenapp/test/media_test.dart
@@ -1,6 +1,3 @@
-import 'package:cogniopenapp/src/media.dart';
-import 'package:flutter_test/flutter_test.dart';
-
 // TODO -- Fix Tests
 
 /*
